requirements.txt
Flask
requests
google-generativeai
concurrent.futures